requirements.txt
asgiref==3.5.2
astroid==2.11.7
autopep8==1.6.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
dill==0.3.5.1
Django==4.0.7
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.9.1
mccabe==0.7.0
oauthlib==3.2.0
platformdirs==2.5.2
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.0
pylint==2.14.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
uritemplate==4.1.1
urllib3==1.26.11
wrapt==1.14.1
